Enum is no longer a problem
I have added a better way of handling the fact that networking package
should not be a part of the dell package, but in a common package. Using
a scriptingfile that sets a suitable PYTHONPATH variable. I have also
done a nice error handeling system. Errors is written to a file if
possible for later debugging.